handle edit compet

+++ InscriptionsSportives/src/gui/controllers/MenuCompetController.java
@@ -26,10 +26,15 @@ import javafx.fxml.Initializable;
 import javafx.scene.Node;
 import javafx.scene.Parent;
 import javafx.scene.Scene;
+import javafx.scene.control.Alert;
+import javafx.scene.control.Alert.AlertType;
+import javafx.scene.control.Button;
 import javafx.scene.control.ListView;
 import javafx.scene.control.TableColumn;
 import javafx.scene.control.TableView;
 import javafx.scene.control.cell.PropertyValueFactory;
+import javafx.scene.layout.BorderPane;
+import javafx.stage.Modality;
 import javafx.stage.Stage;
 
 public class MenuCompetController implements Initializable {
@@ -45,6 +50,10 @@ public class MenuCompetController implements Initializable {
 	private TableColumn<Competition,String> enequipe;
 	@FXML
 	private TableColumn<Competition, String> datecloture;
+	
+	//Pour desactive le bouton quand rien n'est selectionné
+	@FXML 
+	Button DeleteButton = new Button() ; 
 	
 	private String previouslocation="";
 	
@@ -73,16 +82,23 @@ public class MenuCompetController implements Initializable {
 		
 	}
 	
+	//Supprime une compet de l'application
 	@FXML
 	public void handleDeleteCompet(ActionEvent e) throws IOException {
 		
 		Competition competselected= competsTable.getSelectionModel().getSelectedItem() ;
 		
-		for (Candidat candidat : competselected.getCandidats()) {
-			competselected.remove(candidat)	;
+		if(competselected !=null) {
+			for (Candidat candidat : competselected.getCandidats()) {
+				competselected.remove(candidat)	;
+			}
+			competselected.delete();
+			competsTable.getItems().remove(competselected);
 		}
-		competselected.delete();
-		competsTable.getItems().remove(competselected);
+		//rien n'est selectionné en envoie une alert
+		else {
+	       MainController.triggerNoSelectionAlert();
+		}
 		
 	}
 	
@@ -90,13 +106,15 @@ public class MenuCompetController implements Initializable {
 	@Override
 	public void initialize(URL arg0, ResourceBundle arg1) {
 		this.setPreviousLocation(arg0.getFile());
-		
+	// boutton de suppresion desactivé tant qu'un objet n'est pas selectionné
+	//	DeleteButton.setVisible(false);
 		Inscriptions inscriptions =  Inscriptions.getInscriptions();
 		ObservableList<Competition>compets = FXCollections.observableArrayList(inscriptions.getCompetitions());
 		listview.setItems(compets);
 		if(arg0.getFile().endsWith("GestionCompets.fxml")){
 		initializeCompetTable(compets);
 		}
+		
 	
 	}
 
@@ -116,8 +134,13 @@ public class MenuCompetController implements Initializable {
 			datecloture.setCellValueFactory(cellData -> 
 			new ReadOnlyStringWrapper(cellData.getValue().getDateCloture().format(DateTimeFormatter.ofLocalizedDate(FormatStyle.FULL)
 			          .withLocale(Locale.FRENCH))));
+			
 			//association de la table view aux compets
 			competsTable.setItems(compets);
+			
+			// ecoute les changements sur la competTable et affiche les bouttons de suppressions si une compet est selectionnée
+		//    competsTable.getSelectionModel().selectedItemProperty().addListener(
+		  //  (observable, oldValue, newValue) -> DeleteButton.setVisible(true));
 	}
 	
 	@FXML
@@ -137,6 +160,63 @@ public class MenuCompetController implements Initializable {
 		else
 			this.previouslocation="main.fxml";
 		
+	}
+	
+	//Modifie une compet de l'application
+		@FXML
+		public void handleEditCompet(ActionEvent e) throws IOException {
+			
+			Competition competselected= competsTable.getSelectionModel().getSelectedItem() ;
+			if(competselected !=null) {
+				showCompetEditDialog(competselected);
+				this.previouslocation="GestionCompets.fxml";
+				this.backtoMainMenu(e);
+			}
+			else {
+			       MainController.triggerNoSelectionAlert();
+				}
+		}
+		
+		//Ajout un candidat dans la compet
+				@FXML
+				public void handleAddCandidatCompet(ActionEvent e) throws IOException {
+					
+					Competition competselected= competsTable.getSelectionModel().getSelectedItem() ;
+					if(competselected !=null) {
+						// TODO 
+					}
+					else {
+					       MainController.triggerNoSelectionAlert();
+						}
+				}
+	
+	public void showCompetEditDialog(Competition compet) {
+	    try {
+	  
+	        // chargement du fxml de la boite de dialogue
+	        FXMLLoader loader = new FXMLLoader();
+	        loader.setLocation(getClass().getResource("../fxml/EditCompet.fxml"));
+	        BorderPane dialogPage = (BorderPane) loader.load();
+	        // Cree la fenetre de dialogue
+	        Stage dialogStage = new Stage();  
+	        dialogStage.setTitle("Modifier une competition");
+	        dialogStage.initModality(Modality.WINDOW_MODAL);
+	        dialogStage.initOwner(null);
+	        Scene scene = new Scene(dialogPage);
+	        dialogStage.setScene(scene);
+	        // initialisation de la compet en recuperant le controller
+	        EditCompetController controller = loader.getController();
+	        //cela servira pour close la fenetre
+	        controller.setDialogStage(dialogStage); 
+	        controller.loadCompet(compet);
+
+	        // affiche la boite de dialogue
+	        dialogStage.showAndWait();
+	        
+
+	    } catch (IOException e) {
+	        e.printStackTrace();
+	    }
 	}
